Change how the genesis block is included
